fix(前端访问): 隔离过期解锁请求并收紧预览校验

+++ web/src/entities/short-link/api.ts
@@ -39,7 +39,7 @@ const publicShortLinkPreviewSchema: z.ZodType<PublicShortLinkPreview> = z.object
   slug: z.string(),
   targetHost: z.string(),
   redirectMode: z.enum(['direct', 'intermediate']),
-  intermediateDelaySeconds: z.number().min(3).max(10),
+  intermediateDelaySeconds: z.number().int().min(3).max(10),
   expiresAt: z.string().nullable(),
   requiresPassword: z.boolean(),
 })
